requirements.txt
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.2.1
Flask==2.2.2
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-WTF==1.0.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.2
pymongo==4.2.0
python-dateutil==2.8.2
six==1.16.0
Werkzeug==2.2.2
WTForms==3.0.1
gunicorn
python-dotenv==0.21.0
